Make grid component optional
The compiled CSS selectors for the grid component include a total of
1,752 attribute selectors, which severely impact performance on WebKit.
In cases where an application exhibits poor perforance and does not use
the grid, disabling the latter can in some cases result in a big
speedup.

This commit adds a SCSS variable called $include-component-grid, which
is set to true by default. An app can change this to false, e.g. by
including the following in the application-wide .scss file:

    $include-component-grid: false;